Fixed pr comments
